+++ resteasy/src/main/java/com/cerner/beadledom/resteasy/ResteasyContextListener.java
@@ -3,14 +3,13 @@ package com.cerner.beadledom.resteasy;
 import com.cerner.beadledom.lifecycle.GuiceLifecycleContainers;
 import com.cerner.beadledom.lifecycle.LifecycleContainer;
 import com.cerner.beadledom.lifecycle.LifecycleInjector;
-import com.google.common.collect.Lists;
 import com.google.inject.Injector;
 import com.google.inject.Module;
+import java.util.ArrayList;
 import java.util.List;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
-import org.jboss.resteasy.plugins.guice.ModuleProcessor;
 import org.jboss.resteasy.plugins.server.servlet.ResteasyBootstrap;
 import org.jboss.resteasy.spi.Registry;
 import org.jboss.resteasy.spi.ResteasyDeployment;
@@ -37,18 +36,18 @@ public abstract class ResteasyContextListener extends ResteasyBootstrap implemen
         (ResteasyDeployment) context.getAttribute(ResteasyDeployment.class.getName());
     final Registry registry = deployment.getRegistry();
     final ResteasyProviderFactory providerFactory = deployment.getProviderFactory();
-    final ModuleProcessor processor = new ModuleProcessor(registry, providerFactory);
+    final InjectorProcessor processor = new InjectorProcessor(registry, providerFactory);
 
     final List<? extends Module> appModules = getModules(context);
 
-    List<Module> modules = Lists.newArrayList();
+    List<Module> modules = new ArrayList<>();
     modules.addAll(appModules);
 
     injector = GuiceLifecycleContainers.initialize(this, modules);
 
     withInjector(injector);
 
-    processor.processInjector(injector);
+    processor.process(injector);
   }
 
   @Override