+++ resteasy/src/test/java/com/cerner/beadledom/resteasy/TestBindingAnnotation.java
@@ -1,0 +1,21 @@
+package com.cerner.beadledom.resteasy;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import com.google.inject.BindingAnnotation;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+/**
+ * Test binding annotation.
+ *
+ * @author Charan Panchagnula
+ */
+@BindingAnnotation
+@Target({ FIELD, PARAMETER, METHOD })
+@Retention(RUNTIME)
+public @interface TestBindingAnnotation {
+}