+++ resteasy/src/test/java/com/cerner/beadledom/resteasy/TestJaxRsProvider.java
@@ -1,0 +1,15 @@
+package com.cerner.beadledom.resteasy;
+
+import javax.ws.rs.ext.Provider;
+
+/**
+ * JAX-RS provider for testing.
+ *
+ * @author Charan Panchagnula
+ */
+@Provider
+public class TestJaxRsProvider {
+
+  public TestJaxRsProvider() {
+  }
+}