+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * Update maven archetype for project generation with new client recommendations.
 * Update client documentation to recommend using an abstracted resource pattern.
 
+### Fixed
+* JAX-RS providers with binding annotations will no longer be registered in the resteasy provider factory for the server.
+
 ## 3.2.5 - 25 March 2019
 
 ### Defects Corrected